+++ src/tm/assist.ts
@@ -1,9 +1,11 @@
 'use strict';
 
-import { StatusBarAlignment, StatusBarItem, Terminal, window } from "vscode";
+import { basename } from 'path';
+import * as vscode from "vscode";
 import { CommandId, ContextualCommandPalette } from "../commands";
 import { LeafManager, LEAF_EVENT } from "../leaf/core";
 import { LEGATO_ENV } from "../legato/core";
+import { chooseFile, listUpdateFiles } from "../legato/files";
 import { CommandRegister } from '../utils';
 
 const TARGET_SHELL_LABEL = 'Device shell';
@@ -11,19 +13,19 @@ const LOG_SHELL_LABEL = 'Device logs';
 
 export class TargetUiManager extends CommandRegister {
 
-  private targetStatusbar: StatusBarItem;
+  private targetStatusbar: vscode.StatusBarItem;
   private remoteTerminal = new RemoteTerminal(TARGET_SHELL_LABEL, "/bin/sh", ["-c", "ssh root@$DEST_IP"]);
   private logTerminal = new RemoteTerminal(LOG_SHELL_LABEL, "/bin/sh", ["-c", "ssh root@$DEST_IP \"/sbin/logread -f\""]);
   private paletteOnDeviceIP: ContextualCommandPalette;
-  
+
   public constructor() {
     super();
 
     // Status bar
-    this.targetStatusbar = window.createStatusBarItem(StatusBarAlignment.Left, 5);
+    this.targetStatusbar = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 5);
     this.toDispose(this.targetStatusbar); // Dispose status bar on deactivate
     this.targetStatusbar.text = "<Unknown>";
-    this.targetStatusbar.tooltip = "Legato device IP address";
+    this.targetStatusbar.tooltip = "Legato Device";
     this.targetStatusbar.show();
 
     // Commands declaration to be used as QuickPickItem
@@ -33,7 +35,7 @@ export class TargetUiManager extends CommandRegister {
       [
         {
           id: CommandId.SET_DEVICE_IP,
-          label: "Set Device IP",
+          label: "Set Device IP address...",
           callback: () => this.askForNewIP()
         },
         {
@@ -46,8 +48,15 @@ export class TargetUiManager extends CommandRegister {
           label: "Open Device logs",
           callback: () => this.logTerminal.show()
         }
+        ,
+        {
+          id: CommandId.DEVICE_INSTALL_ON,
+          label: "Install app/system on device...",
+          callback: (args:any[]) => this.installOnDevice(args)
+        }
+
       ],
-      'Select command to apply on device...');
+      'Select the command to apply on device...');
     this.paletteOnDeviceIP.register();
 
     // Listen to env changes
@@ -70,13 +79,13 @@ export class TargetUiManager extends CommandRegister {
   private async onEnvVarsChange(oldEnvVars?: any, newEnvVars?: any) {
     let legatoDeviceIpChange = newEnvVars ? newEnvVars[LEGATO_ENV.DEST_IP] : undefined;
     if (legatoDeviceIpChange) {
-      this.updateIPStatusBar(legatoDeviceIpChange);
+      this.targetStatusbar.text = legatoDeviceIpChange;
     }
   }
 
   private async askForNewIP() {
     let ip = await LeafManager.getInstance().getEnvValue(LEGATO_ENV.DEST_IP);
-    let newIP = await window.showInputBox({
+    let newIP = await vscode.window.showInputBox({
       prompt: "Please set the Legato device IP address",
       placeHolder: ip
     });
@@ -86,15 +95,24 @@ export class TargetUiManager extends CommandRegister {
     }
   }
 
-  private async updateIPStatusBar(this: any, ip: string | undefined) {
-    if (ip) {
-      this.targetStatusbar.text = ip;
+  private async installOnDevice(...selectedFile: any[]) {
+    let updateFiles: vscode.Uri[] = await listUpdateFiles();
+    let selectedUpdateFile = selectedFile[0] ? selectedFile[0] :  await chooseFile(updateFiles,
+      {
+        noFileFoundMessage: "No *.update files found in workspace.",
+        quickPickPlaceHolder: "Please select an update file among ones available in the workspace..."
+      });
+
+    if (selectedUpdateFile) {
+      let command = `update ${selectedUpdateFile.path}`;
+      LeafManager.getInstance().taskManager.executeAsTask(`Install ${basename(selectedUpdateFile.path)} on device`, command, await LeafManager.getInstance().getEnvVars());
     }
   }
+
 }
 
-class RemoteTerminal {
-  private leafTerminal: Terminal | undefined;
+export class RemoteTerminal {
+  private leafTerminal: vscode.Terminal | undefined;
   private terminalLabel: string;
   private shellPath: string;
   private shellArgs: string[];
@@ -103,7 +121,7 @@ class RemoteTerminal {
     this.shellPath = path;
     this.shellArgs = shellArgs;
 
-    window.onDidCloseTerminal((closedTerminal: Terminal) => {
+    vscode.window.onDidCloseTerminal((closedTerminal: vscode.Terminal) => {
       if (closedTerminal.name === this.terminalLabel) {
         closedTerminal.dispose();
         this.leafTerminal = undefined;
@@ -113,7 +131,7 @@ class RemoteTerminal {
 
   public async show(preserveFocus?: boolean) {
     if (!this.leafTerminal) {
-      this.leafTerminal = window.createTerminal({
+      this.leafTerminal = vscode.window.createTerminal({
         name: this.terminalLabel,
         shellPath: this.shellPath,
         shellArgs: this.shellArgs,